+++ core-ng/src/main/java/core/framework/impl/db/ConnectionHelper.java
@@ -8,10 +8,10 @@ import java.sql.SQLException;
 /**
  * @author neo
  */
-public class ConnectionTester {
+final class ConnectionHelper {
     // refer to com.mysql.jdbc.integration.c3p0.MysqlConnectionTester, and standard sql state list:
     // http://dev.mysql.com/doc/connector-j/en/connector-j-reference-error-sqlstates.html
-    void checkConnectionStatus(PoolItem<Connection> connection, SQLException e) {
+    static void checkConnectionStatus(PoolItem<Connection> connection, SQLException e) {
         String state = e.getSQLState();
         if (state != null && state.startsWith("08")) {
             connection.broken = true;